added response to file, moved fetch data
The function to fetch the response from API items is
moved to fetch_items.py.
items_to_file.py saves the response to data/response.json.
The idea is to read the data from this file instead of sending
a request to the API.
This file also updates the JSON file if it is older than 15 min.
get_buildid_data is updated to read the data from response.json
instead of the API server.